rename setting folders
