+++ src/Docs/Admin/Controllers/Settings/LocaleController.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace Webkul\RestApi\Docs\Admin\Controllers\Setting;
+namespace Webkul\RestApi\Docs\Admin\Controllers\Settings;
 
-class UserController
+class LocaleController
 {
     /**
      * @OA\Get(
-     *      path="/api/v1/admin/settings/users",
-     *      operationId="getUsers",
-     *      tags={"Users"},
-     *      summary="Get admin users list",
-     *      description="Returns user list, if you want to retrieve all users at once pass pagination=0 otherwise ignore this parameter",
+     *      path="/api/v1/admin/settings/locales",
+     *      operationId="getSettingLocales",
+     *      tags={"Locales"},
+     *      summary="Get admin locale list",
+     *      description="Returns locale list, if you want to retrieve all locales at once pass pagination=0 otherwise ignore this parameter",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="User id",
+     *          description="Locale id",
      *          required=false,
      *          in="query",
      *
@@ -79,7 +79,7 @@ class UserController
      *                  property="data",
      *                  type="array",
      *
-     *                  @OA\Items(ref="#/components/schemas/User")
+     *                  @OA\Items(ref="#/components/schemas/Locale")
      *              ),
      *
      *              @OA\Property(
@@ -96,16 +96,16 @@ class UserController
 
     /**
      * @OA\Get(
-     *      path="/api/v1/admin/settings/users/{id}",
-     *      operationId="getUser",
-     *      tags={"Users"},
-     *      summary="Get admin user detail",
-     *      description="Returns user detail",
+     *      path="/api/v1/admin/settings/locales/{id}",
+     *      operationId="getSalesLocale",
+     *      tags={"Locales"},
+     *      summary="Get admin locale detail",
+     *      description="Returns locale detail",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="User ID",
+     *          description="Locale id",
      *          required=true,
      *          in="path",
      *
@@ -123,7 +123,7 @@ class UserController
      *              @OA\Property(
      *                  property="data",
      *                  type="object",
-     *                  ref="#/components/schemas/User"
+     *                  ref="#/components/schemas/Locale"
      *              )
      *          )
      *      )
@@ -135,11 +135,11 @@ class UserController
 
     /**
      * @OA\Post(
-     *      path="/api/v1/admin/settings/users",
-     *      operationId="storeUsers",
-     *      tags={"Users"},
-     *      summary="Store the user",
-     *      description="Store the user",
+     *      path="/api/v1/admin/settings/locales",
+     *      operationId="storeSettingLocale",
+     *      tags={"Locales"},
+     *      summary="Store the locale",
+     *      description="Store the locale",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\RequestBody(
@@ -150,38 +150,27 @@ class UserController
      *              @OA\Schema(
      *
      *                  @OA\Property(
+     *                      property="code",
+     *                      type="string",
+     *                      description="Language code",
+     *                      example="en"
+     *                  ),
+     *                  @OA\Property(
      *                      property="name",
      *                      type="string",
-     *                      description="User Full Name",
-     *                      example="Peter Doe"
+     *                      example="English"
      *                  ),
      *                  @OA\Property(
-     *                      property="email",
-     *                      type="email",
-     *                      example="peter@example.com"
+     *                      property="direction",
+     *                      type="string",
+     *                      example="ltr",
+     *                      enum={"ltr", "rtl"}
      *                  ),
      *                  @OA\Property(
-     *                      property="password",
-     *                      type="password",
-     *                      example="admin123"
+     *                      property="locale_image[]",
+     *                      type="file"
      *                  ),
-     *                  @OA\Property(
-     *                      property="password_confirmation",
-     *                      type="password",
-     *                      example="admin123"
-     *                  ),
-     *                  @OA\Property(
-     *                      property="status",
-     *                      type="integer",
-     *                      example=1,
-     *                      enum={0, 1}
-     *                  ),
-     *                  @OA\Property(
-     *                      property="role_id",
-     *                      type="integer",
-     *                      example=1
-     *                  ),
-     *                  required={"name", "email", "password", "password_confirmation", "role_id"}
+     *                  required={"code", "name", "direction"}
      *              )
      *          )
      *      ),
@@ -192,8 +181,8 @@ class UserController
      *
      *          @OA\JsonContent(
      *
-     *              @OA\Property(property="message", type="string", example="User created successfully."),
-     *              @OA\Property(property="data", type="object", ref="#/components/schemas/User")
+     *              @OA\Property(property="message", type="string", example="Locale added successfully."),
+     *              @OA\Property(property="data", type="object", ref="#/components/schemas/Locale")
      *          )
      *      ),
      *
@@ -207,7 +196,7 @@ class UserController
      *
      *          @OA\JsonContent(
      *
-     *              @OA\Examples(example="result", value={"message":"The code may not be greater than 3 characters."}, summary="An result object."),
+     *              @OA\Examples(example="result", value={"message":"The code has already been taken."}, summary="An result object."),
      *          )
      *      )
      * )
@@ -217,17 +206,17 @@ class UserController
     }
 
     /**
-     * @OA\Post(
-     *      path="/api/v1/admin/settings/users/{id}",
-     *      operationId="updateUsers",
-     *      tags={"Users"},
-     *      summary="Update user",
-     *      description="Update user",
+     * @OA\Put(
+     *      path="/api/v1/admin/settings/locales/{id}",
+     *      operationId="updateSettingLocale",
+     *      tags={"Locales"},
+     *      summary="Update locale",
+     *      description="Update locale",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="User ID",
+     *          description="Locale id",
      *          required=true,
      *          in="path",
      *
@@ -239,49 +228,27 @@ class UserController
      *      @OA\RequestBody(
      *
      *          @OA\MediaType(
-     *              mediaType="multipart/form-data",
+     *              mediaType="application/json",
      *
      *              @OA\Schema(
      *
      *                  @OA\Property(
-     *                      property="_method",
+     *                      property="code",
      *                      type="string",
-     *                      example="PUT"
+     *                      example="en"
      *                  ),
      *                  @OA\Property(
      *                      property="name",
      *                      type="string",
-     *                      description="User Full Name",
-     *                      example="Peter Doe"
+     *                      example="English"
      *                  ),
      *                  @OA\Property(
-     *                      property="email",
-     *                      type="email",
-     *                      example="peter@example.com"
+     *                      property="direction",
+     *                      type="string",
+     *                      example="ltr",
+     *                      enum={"ltr", "rtl"}
      *                  ),
-     *                  @OA\Property(
-     *                      property="password",
-     *                      type="password",
-     *                      example="admin123"
-     *                  ),
-     *                  @OA\Property(
-     *                      property="password_confirmation",
-     *                      type="password",
-     *                      example="admin123"
-     *                  ),
-     *                  @OA\Property(
-     *                      property="status",
-     *                      type="integer",
-     *                      example=1,
-     *                      enum={0, 1}
-     *                  ),
-     *                  @OA\Property(
-     *                      property="role_id",
-     *                      format="id",
-     *                      type="integer",
-     *                      example=1
-     *                  ),
-     *                  required={"name", "email", "password", "password_confirmation", "role_id"}
+     *                  required={"code", "name", "direction"}
      *              )
      *          )
      *      ),
@@ -295,11 +262,11 @@ class UserController
      *              @OA\Property(
      *                  property="message",
      *                  type="string",
-     *                  example="User updated successfully."),
+     *                  example="Locale updated successfully."),
      *              @OA\Property(
      *                  property="data",
      *                  type="object",
-     *                  ref="#/components/schemas/User"
+     *                  ref="#/components/schemas/Locale"
      *              )
      *          )
      *      ),
@@ -307,6 +274,15 @@ class UserController
      *      @OA\Response(
      *          response=401,
      *          description="Unauthenticated",
+     *      ),
+     *      @OA\Response(
+     *          response=422,
+     *          description="Error: Unprocessable Content",
+     *
+     *          @OA\JsonContent(
+     *
+     *              @OA\Examples(example="result", value={"message":"The code has already been taken."}, summary="An result object."),
+     *          )
      *      )
      * )
      */
@@ -316,16 +292,16 @@ class UserController
 
     /**
      * @OA\Delete(
-     *      path="/api/v1/admin/settings/users/{id}",
-     *      operationId="deleteUser",
-     *      tags={"Users"},
-     *      summary="Delete user by id",
-     *      description="Delete user by id",
+     *      path="/api/v1/admin/settings/locales/{id}",
+     *      operationId="deleteLocale",
+     *      tags={"Locales"},
+     *      summary="Delete locale by id",
+     *      description="Delete locale by id",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="User id",
+     *          description="Locale id",
      *          required=true,
      *          in="path",
      *
@@ -343,9 +319,14 @@ class UserController
      *              @OA\Property(
      *                  property="message",
      *                  type="string",
-     *                  example="Admin deleted successfully."),
+     *                  example="Locale Deleted Successfully."),
      *              )
      *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=404,
+     *          description="Resource Not Found"
      *      )
      * )
      */
