+++ src/Docs/Admin/Controllers/Settings/TaxCategoryController.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace Webkul\RestApi\Docs\Admin\Controllers\Setting;
+namespace Webkul\RestApi\Docs\Admin\Controllers\Settings;
 
-class CurrencyController
+class TaxCategoryController
 {
     /**
      * @OA\Get(
-     *      path="/api/v1/admin/settings/currencies",
-     *      operationId="getSettingCurrencies",
-     *      tags={"Currencies"},
-     *      summary="Get admin currency list",
-     *      description="Returns currency list, if you want to retrieve all currencies at once pass pagination=0 otherwise ignore this parameter",
+     *      path="/api/v1/admin/settings/tax-categories",
+     *      operationId="getTaxCategories",
+     *      tags={"Tax-Categories"},
+     *      summary="Get admin tax category list",
+     *      description="Returns tax category list, if you want to retrieve all tax categories at once pass pagination=0 otherwise ignore this parameter",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Currency id",
+     *          description="Tax Category id",
      *          required=false,
      *          in="query",
      *
@@ -79,7 +79,7 @@ class CurrencyController
      *                  property="data",
      *                  type="array",
      *
-     *                  @OA\Items(ref="#/components/schemas/Currency")
+     *                  @OA\Items(ref="#/components/schemas/TaxCategory")
      *              ),
      *
      *              @OA\Property(
@@ -96,16 +96,16 @@ class CurrencyController
 
     /**
      * @OA\Get(
-     *      path="/api/v1/admin/settings/currencies/{id}",
-     *      operationId="getSalesCurrency",
-     *      tags={"Currencies"},
-     *      summary="Get admin currency detail",
-     *      description="Returns currency detail",
+     *      path="/api/v1/admin/settings/tax-categories/{id}",
+     *      operationId="getTaxCategory",
+     *      tags={"Tax-Categories"},
+     *      summary="Get admin tax category detail",
+     *      description="Returns tax category detail",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Currency ID",
+     *          description="Tax Category id",
      *          required=true,
      *          in="path",
      *
@@ -123,7 +123,7 @@ class CurrencyController
      *              @OA\Property(
      *                  property="data",
      *                  type="object",
-     *                  ref="#/components/schemas/Currency"
+     *                  ref="#/components/schemas/TaxCategory"
      *              )
      *          )
      *      )
@@ -135,39 +135,46 @@ class CurrencyController
 
     /**
      * @OA\Post(
-     *      path="/api/v1/admin/settings/currencies",
-     *      operationId="storeSettingCurrency",
-     *      tags={"Currencies"},
-     *      summary="Store the currency",
-     *      description="Store the currency",
+     *      path="/api/v1/admin/settings/tax-categories",
+     *      operationId="storeTaxCategory",
+     *      tags={"Tax-Categories"},
+     *      summary="Store the tax category",
+     *      description="Store the tax category",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\RequestBody(
      *
      *          @OA\MediaType(
-     *              mediaType="multipart/form-data",
+     *              mediaType="application/json",
      *
      *              @OA\Schema(
      *
      *                  @OA\Property(
      *                      property="code",
      *                      type="string",
-     *                      description="Currency code",
-     *                      example="INR",
-     *                      minLength=3,
-     *                      maxLength=3
+     *                      description="Tax category's code",
+     *                      example="service-tax"
      *                  ),
      *                  @OA\Property(
      *                      property="name",
      *                      type="string",
-     *                      example="Rupee"
+     *                      description="Tax category's name",
+     *                      example="Service Tax"
      *                  ),
      *                  @OA\Property(
-     *                      property="symbol",
+     *                      property="description",
      *                      type="string",
-     *                      example="₹"
+     *                      description="Tax category's description",
+     *                      example="This tax category will apply to all states of a country."
      *                  ),
-     *                  required={"code", "name"}
+     *                  @OA\Property(
+     *                      property="taxrates",
+     *                      description="Tax category's rates",
+     *                      type="array",
+     *
+     *                      @OA\Items(type="integer", example=1)
+     *                  ),
+     *                  required={"code", "name", "description", "taxrates"}
      *              )
      *          )
      *      ),
@@ -178,23 +185,14 @@ class CurrencyController
      *
      *          @OA\JsonContent(
      *
-     *              @OA\Property(property="message", type="string", example="Currency created successfully."),
-     *              @OA\Property(property="data", type="object", ref="#/components/schemas/Currency")
+     *              @OA\Property(property="message", type="string", example="Tax Category created successfully."),
+     *              @OA\Property(property="data", type="object", ref="#/components/schemas/TaxCategory")
      *          )
      *      ),
      *
      *      @OA\Response(
      *          response=401,
      *          description="Unauthenticated",
-     *      ),
-     *      @OA\Response(
-     *          response=422,
-     *          description="Error: Unprocessable Content",
-     *
-     *          @OA\JsonContent(
-     *
-     *              @OA\Examples(example="result", value={"message":"The code may not be greater than 3 characters."}, summary="An result object."),
-     *          )
      *      )
      * )
      */
@@ -204,16 +202,16 @@ class CurrencyController
 
     /**
      * @OA\Put(
-     *      path="/api/v1/admin/settings/currencies/{id}",
-     *      operationId="updateSettingCurrency",
-     *      tags={"Currencies"},
-     *      summary="Update currency",
-     *      description="Update currency",
+     *      path="/api/v1/admin/settings/tax-categories/{id}",
+     *      operationId="updateTaxCategory",
+     *      tags={"Tax-Categories"},
+     *      summary="Update tax category",
+     *      description="Update tax category",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Currency ID",
+     *          description="Tax Category id",
      *          required=true,
      *          in="path",
      *
@@ -232,19 +230,29 @@ class CurrencyController
      *                  @OA\Property(
      *                      property="code",
      *                      type="string",
-     *                      example="INR"
+     *                      description="Tax category's code",
+     *                      example="service-tax"
      *                  ),
      *                  @OA\Property(
      *                      property="name",
      *                      type="string",
-     *                      example="Rupee"
+     *                      description="Tax category's name",
+     *                      example="Service Tax"
      *                  ),
      *                  @OA\Property(
-     *                      property="symbol",
+     *                      property="description",
      *                      type="string",
-     *                      example="₹"
+     *                      description="Tax category's description",
+     *                      example="This tax category will apply to all states of a country."
      *                  ),
-     *                  required={"code", "name", "symbol"}
+     *                  @OA\Property(
+     *                      property="taxrates",
+     *                      description="Tax category's rates",
+     *                      type="array",
+     *
+     *                      @OA\Items(type="integer", example=1)
+     *                  ),
+     *                  required={"code", "name", "description", "taxrates"}
      *              )
      *          )
      *      ),
@@ -258,11 +266,11 @@ class CurrencyController
      *              @OA\Property(
      *                  property="message",
      *                  type="string",
-     *                  example="Currency updated successfully."),
+     *                  example="Tax Category updated successfully."),
      *              @OA\Property(
      *                  property="data",
      *                  type="object",
-     *                  ref="#/components/schemas/Currency"
+     *                  ref="#/components/schemas/TaxCategory"
      *              )
      *          )
      *      ),
@@ -270,15 +278,6 @@ class CurrencyController
      *      @OA\Response(
      *          response=401,
      *          description="Unauthenticated",
-     *      ),
-     *      @OA\Response(
-     *          response=422,
-     *          description="Error: Unprocessable Content",
-     *
-     *          @OA\JsonContent(
-     *
-     *              @OA\Examples(example="result", value={"message":"The code has already been taken."}, summary="An result object."),
-     *          )
      *      )
      * )
      */
@@ -288,16 +287,16 @@ class CurrencyController
 
     /**
      * @OA\Delete(
-     *      path="/api/v1/admin/settings/currencies/{id}",
-     *      operationId="deleteCurrency",
-     *      tags={"Currencies"},
-     *      summary="Delete currency by id",
-     *      description="Delete currency by id",
+     *      path="/api/v1/admin/settings/tax-categories/{id}",
+     *      operationId="deleteTaxCategory",
+     *      tags={"Tax-Categories"},
+     *      summary="Delete tax category by id",
+     *      description="Delete tax category by id",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Currency id",
+     *          description="Tax Category id",
      *          required=true,
      *          in="path",
      *
@@ -315,7 +314,7 @@ class CurrencyController
      *              @OA\Property(
      *                  property="message",
      *                  type="string",
-     *                  example="Currency Deleted Successfully."),
+     *                  example="Tax Category deleted successfully."),
      *              )
      *          )
      *      )
