+++ src/Http/Controllers/V1/Admin/Settings/ChannelController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Event;
 use Webkul\Core\Repositories\ChannelRepository;
 use Webkul\Core\Rules\Code;
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\ChannelResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\ChannelResource;
 
 class ChannelController extends SettingController
 {
