+++ src/Http/Controllers/V1/Admin/Settings/UserController.php
@@ -5,7 +5,7 @@ namespace Webkul\RestApi\Http\Controllers\V1\Admin\Settings;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Str;
 use Webkul\Admin\Http\Requests\UserForm;
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\UserResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\UserResource;
 use Webkul\User\Repositories\AdminRepository;
 
 class UserController extends SettingController
