+++ src/Http/Resources/V1/Admin/Settings/CountryResource.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Webkul\RestApi\Http\Resources\V1\Admin\Setting;
+namespace Webkul\RestApi\Http\Resources\V1\Admin\Settings;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class ThemeResource extends JsonResource
+class CountryResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
@@ -15,7 +15,9 @@ class ThemeResource extends JsonResource
     public function toArray($request)
     {
         return [
-            ...$this->resource->toArray(),
+            'id'   => $this->id,
+            'code' => $this->code,
+            'name' => $this->name,
         ];
     }
 }
