+++ src/Http/Resources/V1/Admin/Settings/LocaleResource.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Webkul\RestApi\Http\Resources\V1\Admin\Setting;
+namespace Webkul\RestApi\Http\Resources\V1\Admin\Settings;
 
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class UserResource extends JsonResource
+class LocaleResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
@@ -16,10 +16,8 @@ class UserResource extends JsonResource
     {
         return [
             'id'         => $this->id,
+            'code'       => $this->code,
             'name'       => $this->name,
-            'email'      => $this->email,
-            'status'     => $this->status,
-            'role'       => new RoleResource($this->role),
             'created_at' => $this->created_at,
             'updated_at' => $this->updated_at,
         ];
