+++ src/Http/Resources/V1/Shop/Core/ChannelResource.php
@@ -2,7 +2,7 @@
 
 namespace Webkul\RestApi\Http\Resources\V1\Shop\Core;
 
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\ChannelResource as AdminChannelResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\ChannelResource as AdminChannelResource;
 
 class ChannelResource extends AdminChannelResource
 {
