+++ src/Http/Resources/V1/Shop/Core/ConfigurationResource.php
@@ -2,7 +2,7 @@
 
 namespace Webkul\RestApi\Http\Resources\V1\Shop\Core;
 
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\ConfigurationResource as AdminConfigurationResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\ConfigurationResource as AdminConfigurationResource;
 
 class ConfigurationResource extends AdminConfigurationResource
 {
