+++ src/Http/Resources/V1/Shop/Core/CountryResource.php
@@ -2,7 +2,7 @@
 
 namespace Webkul\RestApi\Http\Resources\V1\Shop\Core;
 
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\CountryResource as AdminCountryResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\CountryResource as AdminCountryResource;
 
 class CountryResource extends AdminCountryResource
 {
