+++ src/Http/Resources/V1/Shop/Core/CountryStateResource.php
@@ -2,7 +2,7 @@
 
 namespace Webkul\RestApi\Http\Resources\V1\Shop\Core;
 
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\CountryStateResource as AdminCountryStateResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\CountryStateResource as AdminCountryStateResource;
 
 class CountryStateResource extends AdminCountryStateResource
 {
