+++ src/Http/Resources/V1/Shop/Core/CurrencyResource.php
@@ -2,7 +2,7 @@
 
 namespace Webkul\RestApi\Http\Resources\V1\Shop\Core;
 
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\CurrencyResource as AdminCurrencyResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\CurrencyResource as AdminCurrencyResource;
 
 class CurrencyResource extends AdminCurrencyResource
 {
