+++ src/Http/Resources/V1/Shop/Core/LocaleResource.php
@@ -2,7 +2,7 @@
 
 namespace Webkul\RestApi\Http\Resources\V1\Shop\Core;
 
-use Webkul\RestApi\Http\Resources\V1\Admin\Setting\LocaleResource as AdminLocaleResource;
+use Webkul\RestApi\Http\Resources\V1\Admin\Settings\LocaleResource as AdminLocaleResource;
 
 class LocaleResource extends AdminLocaleResource
 {
